fix float properties in abstract exporter
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk@4805 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ jasperreports/src/net/sf/jasperreports/engine/JRAbstractExporter.java
@@ -259,11 +259,11 @@ public abstract class JRAbstractExporter implements JRExporter
 				Float floatValue = (Float)parameters.get(parameter);
 				if (floatValue == null)
 				{
-					return JRProperties.getIntegerProperty(property);
+					return JRProperties.getFloatProperty(property);
 				}
 				else
 				{
-					return floatValue.intValue();
+					return floatValue.floatValue();
 				}
 			}
 			else
